add more response class and update mapper
