+++ auctionhouse/src/main/java/com/weareadaptive/auction/controller/dto/BidInfo.java
@@ -1,0 +1,8 @@
+package com.weareadaptive.auction.controller.dto;
+
+import com.weareadaptive.auction.model.Bid;
+
+public record BidInfo(String bidder, int quantity, double price, Bid.State state,
+                      int winQuantity
+) {
+}
